Refactor GetActiveApp() into platform package, do app.down() in cleanup
