+++ cmd/ddev/cmd/restart_test.go
@@ -19,7 +19,7 @@ func TestDevRestart(t *testing.T) {
 		out, err := exec.RunCommand(DdevBin, args)
 		assert.NoError(err)
 
-		app, err := getActiveApp("")
+		app, err := platform.GetActiveApp("")
 		if err != nil {
 			assert.Fail("Could not find an active ddev configuration: %v", err)
 		}
